# book-social-network/book-network/src/main/java/com/daniel/book/feedback/FeedbackResponse.java
package com.daniel.book.feedback;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class FeedbackResponse {

    private Double note;
    private String comment;
    private boolean ownFeedback;
}